drivers: clock: Implement system clock mux controller for CH56x
The CH56x can be configured to run on a 30MHz external oscillator or a
480MHz PLL generated by the USB PHY.

This commit introduces a clock controller that can automatically select
one of the two available clock sources and determine the appropriate
divisor based on the given frequency.

The `CONFIG_SYS_CLOCK_HW_CYCLES_PER_SEC` value is also updated with the
selected frequency.

Signed-off-by: Chen Xingyu <hi@xingrz.me>

# Conflicts:
#	drivers/clock_control/Kconfig
#	soc/riscv/riscv-privileged/wch_ch32v/Kconfig.defconfig.ch56x