pm: policy: fix latency request updates
The latency requests made to the power management subsystem is currently
buggy. The main issue is that the latency values are treated as int32_ts
rather than uint32_ts. By casting the latency request as an int32_t the
max latency request gets stuck at -1 because new_max_latency_us is
initialized to SYS_FOREVER_US which is defined as -1. Any requests that
come in which are lesser than INT32_MAX will be ignored since they are not
less than -1. This addresses this problem by converting the types of the
latency variables to uint32_t which should be the case to begin with since
latency requests can only be positive values.

Signed-off-by: Siddharth Gupta <siddharth.gupta2694@gmail.com>