build: namespace the generated headers with `zephyr/`
Namespaced the generated headers with `zephyr` to prevent
potential conflict with other headers.

Introduce a temporary Kconfig `LEGACY_GENERATED_INCLUDE_PATH`
that is deprecated and enabled by default. This allows the
developers to continue the use of the old include paths for the
time being until it is removed. The Kconfig will generate a
build-time deprecation message.

Updated the includes path of in-tree sources accordingly.

Most of the changes here are scripted, check the PR for more info.

Signed-off-by: Yong Cong Sin <ycsin@meta.com>