sca: Introduce SCA Kconfig
Now the integration accepts a kconfig file for configuring the
analysis and the generation of the reports. The path of the
kconfig file should be provided via the variable ECLAIR_CONFIG.

Without providing the kconfig file the first_analysis set will
be used and no additional artefacts other than the database will
be produced.

Avoidable uses of getenv in the ECL files have been removed:
now the options are provided directly in the eclair_(env|report)
invocations in sca.cmake.

reports.ecl is not supposed anymore to produce reports, now
this task is handled directly by sca.cmake. The file has been
renamed in db_generation.ecl as this is its only task.

Signed-off-by: Simone Ballarin <simone.ballarin@bugseng.com>