debug: thread_analyzer: Option to analyze threads on each core separately
Add implementation to analyze thread on each core separately. This
feature can be enabled with THREAD_ANALYZER_AUTO_SEPARATE_CORES config
option. If enabled, an analyzer thread is started for each core, and
the threads will only analyze thread on the core its running on.

This feature is needed for Intel ADSP platform, where core specific
caches are not synchronized between the cores.

Signed-off-by: Jyri Sarha <jyri.sarha@linux.intel.com>