kernel: Update IPI usage in k_thread_priority_set()
1. The flagging of IPIs is moved out of k_thread_priority_set() into
z_thread_prio_set(). This allows for an IPI to be done for a thread
that had its priority bumped due to the handling of priority
inheritance from a mutex.

2. Not issuing a call to z_reschedule_unlocked() when the scheduler
is locked is insufficient. If the thread whose priority has changed
is executing on another CPU, then it needs any pending IPIs that have
been previously flagged by z_thread_prio_set() to be signalled.

Incidentally, nearly all calls to flag_ipi() are now performed with
sched_spinlock being locked. The only exception is in slice_timeout().

Signed-off-by: Peter Mitsis <peter.mitsis@intel.com>