kernel: Add CONFIG_ARCH_HAS_DIRECTED_IPIS
Some platforms allow IPIs to be directed to specific CPUs, while
others will broadcast the IPIs to all CPUs. As the kernel has the
capability to track which CPUs may need an IPI (see
CONFIG_IPI_OPTIMIZE), this commit updates the arch_sched_ipi() API
to accept a bitmap of CPUs to which IPIs will be sent.

Platforms that allow directed IPIs may see a significant reduction
in the number of IPI related ISRs when CONFIG_IPI_OPTIMIZE is
enabled and the number of CPUs increases.  These platforms can be
identified by the Kconfig option CONFIG_ARCH_HAS_DIRECTED_IPIS.

Signed-off-by: Peter Mitsis <peter.mitsis@intel.com>