storage: non volatile memory circular buffer
A new circular buffer mechanism (nvcb) is proposed for storing data on
non volatile memory. The solution stores data as cbor binary string
items. The non volatile memory is divided into configurable block sizes.
Data is written sequentially to a block and when a block is full a new
block can be taken into use. When the non volatile memory reserved for
the circular buffer is full the first block is reused to store data.

Reading of data can be performed in a forward method (stepping over
data in increasing blocks), but it can also use a backward method
(stepping forward in a block but backwards in blocks) to allow for
faster retrieving of later added data.

More information on how to set it up can be found in `nvcb.h`.

In comparison with fcb the proposal provides the following:
a. Support for non flash backends, e.g. also eeproms can be used,
b. Content can easily be generated (writing cbor entries),
c. Provide backward stepping over blocks,
d. Support backends with non constant block size,

Backend configuration for flash and eeprom will follow shortly

This backend will be the base of a new key-value store that allows
direct storage and retrieval of key-value items and maintains their
validity when new items are added.

This new key-value store will then be added as a backend for
settings.

Added support for 0 byte length cbor binary strings
Added support for spare blocks configuration (spare blocks can contain
data but are no longer considered to be active/valid)

Signed-off-by: Laczen JMS <laczenjms@gmail.com>