boards: arduino: portenta_h7: Updated Flash Partition Map
I made some changes to the flash partition structure as I eventually
want to support MCU Boot on the portenta H7. To achieve that I first
need to support the power regulator the portenta uses as it needs some
I2C commands on power on to enable the various voltage rails.
However, the old parition structure had only a code partition and I
have rearranged it to include two slots for the images, the bootloader
and scratch to be located at the beginning of the flash so that
it remains compatible with the arduino bootloader if necessary.

Signed-off-by: Rahul Arasikere <arasikere.rahul@gmail.com>