arch: arm: Add hook called when exiting CPU idle state
Added a hook which is called when CPU exits cpu idle. There is already
one optional which is called on entering CPU idle. Those two can be
used to track CPU activity.

There is already a macro hook which is called immediately after the
wake up and it is used to inject specific code where primary use has
was HW errata. This hook is more generic since it is a C function which
user can implement.

Signed-off-by: Krzysztof Chruściński <krzysztof.chruscinski@nordicsemi.no>