drivers: sensor: tmag5273: add support for power management
TMAG5273 supports a low power (5 nA) sleep mode which we can utilize
to save power when not actively sampling.

The chip can be woken up from sleep mode using the INT pin or an I2C
transaction. The decision whether the INT pin or I2C should be used for
wakeup is independent from the decision which method to use for single
conversion triggering. It seems however unlikely that somebody will
want to use one method for wakeup and another method for conversion
triggering. Therefore only one dt property is provided to configure if
the INT pin should be used for triggering in general.

When run-time power management is enabled the resulting behavior of the
device will be similar to the Sleep Mode as described in the data sheet.
It doesn't make a big difference which mode is configured in device tree.
When an application wants to obtain the behavior of Stand-by (Trigger)
Mode or Continous Mode as described in data sheet it needs to make sure
to run pm_device_runtime_get() to keep the device active.

Signed-off-by: Oliver Barta <oliver.barta@grandcentrix.net>