drivers: sensor: tmag5273: fix handling of INT GPIO
The interrupt line is effectively used in open drain mode. The driver is
switching between output low and input.

At the moment the pin is used exclusively for signaling towards the
sensor. Therefore the pin could be configured open drain right away at
the moment. That would avoid the need for the explicit switching between
input and output. I assume that it was implemented this way because it
was expected that the driver will be extended at some point in future to
support also signaling from sensor to host. That seems indeed likely,
therefore I don't want to change the way how it is implemented.

The implementation of the switching between input and output was however
wrong. In order to achieve open drain behavior, it needs to be made sure
that the pin is never actively driven high. As the pin is active low,
we must never do gpio_pin_set_dt(&drv_cfg->int_gpio, 0).

This got a practical problem while adding support for power management.
It turned out that the sensor is driving the INT line low for more than
40 us when waking up from sleep mode. When the tmag5273_dev_int_trigger()
was used to wake up the sensor, the time periods where the pin was
actively driven high on host side and low on sensor side overlapped.

Signed-off-by: Oliver Barta <oliver.barta@grandcentrix.net>