net: ethernet: Add support for embedding ll header
If Ethernet based network device driver is advertizing
ETHERNET_EMBEDDED_LL_HEADER, then the L2 driver will record
the size of the link layer header to the network interface.
This info is then used in TX to determine whether to have a
separate net_buf for the link level header, or to embed the
header to the same net_buf as the L2 payload.
This functionality is only used if a separate Kconfig option
CONFIG_NET_PKT_ALLOW_HEADROOM_ALLOCATION is enabled.

Signed-off-by: Jukka Rissanen <jukka.rissanen@nordicsemi.no>