scripts: runners: Add cvi-fiptool
The SOPHGO CVI series SoCs load the RTOS kernel for the 2nd core from a
`fip.bin` located in the main storage.

This commit introduced a convenient script for updating the `fip.bin` on
the external storage mounted to the host, with the newly built `zephyr.bin`.

Path to the `fiptool.py` from the official SDK should be provided.

Signed-off-by: Chen Xingyu <hi@xingrz.me>