pm: policy: add next event notification
The pm subsystem exposes a way for drivers to register events which allow
the pm policy to make decisions about the next power state of the device.
However there is no way for custom policies to get notified when the next
event will happen. This change creates a subsrciption list similar to the
max latency list which allows subscribers to be notified when the next
event time is updated.

Signed-off-by: Siddharth Gupta <siddharth.gupta2694@gmail.com>